Change test script

+++ test_single.sh
@@ -1,18 +1,21 @@
 #!/bin/bash
 
+# Author : James Nock (@Jpnock)
+# Year   : 2023
+
 set -uo pipefail
+shopt -s globstar
 
-# Clean bin directory
-echo "Cleaning bin directory..."
-rm -rf ./bin/*
-mkdir -p bin/output
-
-# Compile the project
 make bin/c_compiler
 
-# Initialize variables
+rm -rf bin/output
+
+mkdir -p bin
+mkdir -p bin/output
+
 TOTAL=0
 PASSING=0
+
 J_UNIT_OUTPUT_FILE="./bin/junit_results.xml"
 printf '%s\n' '<?xml version="1.0" encoding="UTF-8"?>' > "${J_UNIT_OUTPUT_FILE}"
 printf '%s\n' '<testsuite name="Integration test">' >> "${J_UNIT_OUTPUT_FILE}"
@@ -23,16 +26,16 @@ fail_testcase() {
     printf '%s\n' "</testcase>" >> "${J_UNIT_OUTPUT_FILE}"
 }
 
-test_file=$1 # Take the test name as a command-line argument
+test_folder="custom_tests"
+test_path="$1" # The first command-line argument should be the relative path to the test.
 
-# Remove any trailing .c if present
-test_file="${test_file%.c}"
+# Construct paths for the test and driver files
+TO_ASSEMBLE="${test_folder}/${test_path}.c"
+DRIVER="${test_folder}/${test_path}_driver.c"
+LOG_PATH="${test_path//\//_}"
+LOG_PATH="./bin/output/${LOG_PATH}"
 
-TO_ASSEMBLE="custom_tests/${test_file}.c"
-LOG_PATH="./bin/output/${test_file}"
-OUT="${LOG_PATH}"
-DRIVER="custom_tests/${test_file}_driver.c"
-
+(( TOTAL++ ))
 echo "${TO_ASSEMBLE}"
 printf '%s\n' "<testcase name=\"${TO_ASSEMBLE}\">" >> "${J_UNIT_OUTPUT_FILE}"
 
@@ -47,25 +50,35 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
-riscv64-unknown-elf-gcc -march=rv32imfd -mabi=ilp32d -o "${OUT}.o" -c "${OUT}.s" 2> "${LOG_PATH}.assembler.stderr.log" > "${LOG_PATH}.assembler.stdout.log"
-if [ $? -ne 0 ]; then
-    fail_testcase "Fail: see ${LOG_PATH}.assembler.stderr.log and ${LOG_PATH}.assembler.stdout.log"
-    exit 1
-fi
+# Compile only if driver exists
+if [ -f "${DRIVER}" ]; then
+    riscv64-unknown-elf-gcc -march=rv32imfd -mabi=ilp32d -o "${OUT}.o" -c "${OUT}.s" 2> "${LOG_PATH}.assembler.stderr.log" > "${LOG_PATH}.assembler.stdout.log"
 
-riscv64-unknown-elf-gcc -march=rv32imfd -mabi=ilp32d -static -o "${OUT}" "${OUT}.o" "${DRIVER}" 2> "${LOG_PATH}.linker.stderr.log" > "${LOG_PATH}.linker.stdout.log"
-if [ $? -ne 0 ]; then
-    fail_testcase "Fail: see ${LOG_PATH}.linker.stderr.log and ${LOG_PATH}.linker.stdout.log"
-    exit 1
-fi
+    if [ $? -ne 0 ]; then
+        fail_testcase "Fail: see ${LOG_PATH}.assembler.stderr.log and ${LOG_PATH}.assembler.stdout.log"
+        exit 1
+    fi
 
-spike pk "${OUT}" > "${LOG_PATH}.simulation.log"
-if [ $? -eq 0 ]; then
-    echo -e "\t> Pass"
-    printf '%s\n' "</testcase>" >> "${J_UNIT_OUTPUT_FILE}"
+    riscv64-unknown-elf-gcc -march=rv32imfd -mabi=ilp32d -static -o "${OUT}" "${OUT}.o" "${DRIVER}" 2> "${LOG_PATH}.linker.stderr.log" > "${LOG_PATH}.linker.stdout.log"
+
+    if [ $? -ne 0 ]; then
+        fail_testcase "Fail: see ${LOG_PATH}.linker.stderr.log and ${LOG_PATH}.linker.stdout.log"
+        exit 1
+    fi
+
+    spike pk "${OUT}" > "${LOG_PATH}.simulation.log"
+
+    if [ $? -eq 0 ]; then
+        echo -e "\t> Pass"
+        (( PASSING++ ))
+        printf '%s\n' "</testcase>" >> "${J_UNIT_OUTPUT_FILE}"
+    else
+        fail_testcase "Fail: simulation did not exit with exit-code 0"
+        exit 1
+    fi
 else
-    fail_testcase "Fail: simulation did not exit with exit-code 0"
-    exit 1
+    echo "No driver found. Skipping linking and simulating."
 fi
 
+printf "\nPassing %d/%d tests\n" "${PASSING}" "${TOTAL}"
 printf '%s\n' '</testsuite>' >> "${J_UNIT_OUTPUT_FILE}"
